Feature/character collection

+++ tarantinoverse/View/ObjectView.cpp
@@ -1,7 +1,6 @@
 // ObjectView.cpp
 #include "ObjectView.h"
 #include "View/CharacterView.h"
-
 #include <godot_cpp/classes/area2d.hpp>
 #include <godot_cpp/classes/resource_loader.hpp>
 #include <godot_cpp/classes/sprite2d.hpp>
@@ -79,7 +78,7 @@ void ObjectView::_process(double delta) {
 void ObjectView::_on_body_entered(Node *body) {
   CharacterView *character_view = Object::cast_to<CharacterView>(body);
   if (character_view && presenter) {
-    character_view->collect_object(static_cast<int>(presenter->get_type()),
+    character_view->collect_object(presenter->get_type(),
                                    presenter->get_value());
     presenter->collect();
   }
